% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rpstreams.R
\name{pstream}
\alias{pstream}
\alias{pstream_close}
\alias{send_eof}
\alias{signal}
\title{Open a pipe stream object}
\usage{
pstream(command, args = "")

pstream_close(stream)

send_eof(stream)

signal(stream, signal = 15, group = FALSE)
}
\arguments{
\item{command}{the program to run}

\item{args}{a vector of argument strings}

\item{stream}{a pstream object}

\item{signal}{the POSIX signal number (see \code{kill -l})}

\item{group}{signal the entire process group?}
}
\description{
Opens a set of pipes to stdin, stdout and stderr for
on the specified process.
}
\details{
Closing a stream will wait until the spawned process completes. This
can hang your session if the process is not well-behaved. You should manually
end the proces if possible. \code{pstream_close} will check whether the program
has exited. If it is still running it, EOF is sent. The process is then
checked for a predetermined number of seconds (set by the compile flag KILL_WAIT_SECONDS).
If the process does not exit during that perio, then SIGTERM signal is sent. If
after another round of waiting, the process has not existed, it is then sent
the SIGKILL signal. After that, the stream is manually closed.
}
\examples{
x = pstream("R")
status(x)
read_stderr(x)
pstream_close(x)

x = pstream("R", "--vanilla")
read_stdout(x)
write_stdin(x, "R.Version()")
read_stdout(x)
send_eof(x); Sys.sleep(1)
status(x)

x = pstream("R", "--vanilla")
status(x)
signal(x); Sys.sleep(1)
status(x)

x = pstream("R", "--vanilla")
status(x)
pstream_close(x)
status(x)
}

