% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rpstreams.R
\docType{package}
\name{rpstreams}
\alias{rpstreams}
\alias{rpstreams-package}
\title{Bidirectional, non-blocking communication over pipes}
\description{
The \code{rpstreams} package wraps Jonathan Wakely's pstreams library, which
allows communicating with a subprocess using stdin, stdout and stderr. All
reads are non-blocking.
}
\details{
The \code{\link{pstream}} function opens a stream object linked to a specified
command (a program callable from the command line). Three functions:
\code{\link{write_stdin}}, \code{\link{read_stdout}} and \code{\link{read_stderr}}
can then be used to communicate with the launched proces. \code{\link{status}} will
return information about the state of the stream object.
}
\author{
Timothy H. Keitt

Jonathan Wakeley (pstreams C++ library)
}

