% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rpstreams.R
\name{is_open}
\alias{errno}
\alias{exit_code}
\alias{has_exited}
\alias{is_bad}
\alias{is_eof}
\alias{is_fail}
\alias{is_good}
\alias{is_open}
\alias{status}
\title{Report state of pstream}
\usage{
is_open(stream)

is_eof(stream)

is_good(stream)

is_bad(stream)

is_fail(stream)

has_exited(stream)

exit_code(stream)

errno(stream)

status(stream)
}
\arguments{
\item{stream}{a pstream object}
}
\value{
a boolean
}
\description{
Report state of pstream
}
\examples{
s = pstream("R", "--vanilla")
status(s)
is_open(s)
is_eof(s)
is_good(s)
is_bad(s)
is_fail(s)
write_stdin(s, "q()")
Sys.sleep(1)
has_exited(s)
exit_code(s)
errno(s)
status(s)
}

