% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rpstreams.R
\name{write_stdin}
\alias{read_stderr}
\alias{read_stdout}
\alias{write_stdin}
\title{Read and write data to process}
\usage{
write_stdin(stream, data, send_eof = FALSE)

read_stdout(stream)

read_stderr(stream)
}
\arguments{
\item{stream}{a pstream object}

\item{data}{a vector of values}

\item{send_eof}{if true, write EOF to stream

# Note: it appears that reading from stdout and stderr
# does not work in examples. I suspect those are redirected
# by R internally when running examples.
x = pstream("R")
read_stderr(x)

x = pstream("R", "--vanilla")
read_stdout(x)
write_stdin(x, "R.Version()")
read_stdout(x)
write_stdin(x, "q()")}
}
\description{
Read and write data to process
}

