% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rpstreams.R
\name{\%<<\%}
\alias{\%<<\%}
\alias{\%<>\%}
\alias{\%>>\%}
\title{Infix stream operators}
\usage{
lhs \%<<\% rhs

lhs \%>>\% rhs

lhs \%<>\% rhs
}
\arguments{
\item{lhs}{a pstream object}

\item{rhs}{a data value or name}
}
\value{
a pstream object invisibly
}
\description{
Infix stream operators
}
\details{
The \code{\%<<\%} operator writes the object on the right-hand
side to the pstream object on the left-hand side using
\code{\link{write_stdin}}. The \code{\%>>\%} opertor reads from
the pstream object on the left-hand side and stores the result
in the variable on the right-hand side using \code{\link{read_stdout}}.
The \code{\%<>\%} writes a string to the stream and immediately
reads from standard output and returns the result. This is mostly
for quickly viewing the output from a program.
}
\examples{
x = pstream("R", "--vanilla --slave")
x \%<<\% "R.Version()"
x \%>>\% y
show(y)
x \%<<\% "R.Version()" \%<<\% "dir()" \%>>\% y
show(y)
x \%<>\% "R.Version()"
}

