% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rpstreams.R
\name{\%<<\%}
\alias{\%<<\%}
\alias{\%<>\%}
\alias{\%>>\%}
\title{Infix stream operators}
\usage{
lhs \%<<\% rhs

lhs \%>>\% rhs

lhs \%<>\% rhs
}
\arguments{
\item{lhs}{a pstream object}

\item{rhs}{a data value or name}
}
\value{
a pstream object invisibly
}
\description{
Infix stream operators
}
\examples{
x = pstream("R", "--vanilla --slave")
x \%<<\% "R.Version()"
x \%>>\% y
show(y)
x \%<<\% "R.Version()" \%<<\% "dir()" \%>>\% y
show(y)
x \%<>\% "R.Version()"
}

